// src/modules/schema/types.ts
export interface ErrorTest {
  class?: string
  method?: string
  type: 'invalid param' | 'invalid value' | 'missing value' | 'missing key'
  name: string
  expect: string
  received: any
  index?: number
  message: string
}

export interface SuccessTest {
  class?: string
  method?: string
  name: string
  expect: string
  index?: number
  received: any
}